chore: refactor structure to make it more explicit what is which thing, reducing cognitive load
