+++ src/components/decision/index.js
