+++ src/components/decision/pq.js
