+++ src/index.js
@@ -3,6 +3,7 @@
 const series = require('async/series')
 const retry = require('async/retry')
 const debug = require('debug')
+
 const log = debug('bitswap')
 log.error = debug('bitswap:error')
 const EventEmitter = require('events').EventEmitter
@@ -12,12 +13,12 @@ const paramap = require('pull-paramap')
 const defer = require('pull-defer/source')
 const Block = require('ipfs-block')
 
-const cs = require('./constants')
-const WantManager = require('./wantmanager')
-const Network = require('./network')
-const decision = require('./decision')
+const CONSTANTS = require('./constants')
+const WantManager = require('./components/want-manager')
+const Network = require('./components/network')
+const decision = require('./components/decision')
 
-module.exports = class Bitwap {
+class Bitswap {
   constructor (id, libp2p, blockstore, peerBook) {
     // the ID of the peer to act on behalf of
     this.self = id
@@ -38,7 +39,7 @@ module.exports = class Bitwap {
     this.dupDataRecvd = 0
 
     this.notifications = new EventEmitter()
-    this.notifications.setMaxListeners(cs.maxListeners)
+    this.notifications.setMaxListeners(CONSTANTS.maxListeners)
   }
 
   // handle messages received through the network
@@ -326,14 +327,16 @@ module.exports = class Bitwap {
 }
 
 // Helper method, to add a cid to a block before storing it in the ipfs-repo/blockstore
-function blockToStore (b, cb) {
+function blockToStore (b, callback) {
   b.key((err, key) => {
     if (err) {
-      return cb(err)
+      return callback(err)
     }
-    cb(null, {
+    callback(null, {
       data: b.data,
       key: key
     })
   })
 }
+
+module.exports = Bitswap
