+++ src/types/message/entry.js
