+++ test/wantmanager/index.spec.js
@@ -10,8 +10,8 @@ const Block = require('ipfs-block')
 const CID = require('cids')
 
 const cs = require('../../src/constants')
-const Message = require('../../src/message')
-const WantManager = require('../../src/wantmanager')
+const Message = require('../../src/types/message')
+const WantManager = require('../../src/components/want-manager')
 
 const mockNetwork = require('../utils').mockNetwork
